Move Lightning Address into Spark Settings
Extract lightning_address_ux (claim form, edit form, full display +
confirmation modal) out of view/connect/mod.rs into a dedicated
view/spark/settings/lightning_address.rs module, alongside the
LN_ADDRESS_DOMAIN constant. Convert view/spark/settings.rs into a
directory module so the LN page sits beside the General page.

App::view special-cases Menu::Spark(Settings(Some(LightningAddress)))
to render the LN form with a borrow of ConnectCubePanel — the State
trait's view signature can't reach Connect state, so the dispatch lives
where both panels are owned. The per-Cube Connect → Lightning Address
nav entry still routes to the same lifted function for A/B
verification; that entry point is deleted in the per-Cube Connect
surface removal commit.
